[LongNT] Refactor code
